requirements.txt
streamlit
python-dotenv
Pillow